brain_transform/requirements.txt
pandas
matplotlib
scipy
scikit-learn
dvc[s3]
gitpython
apscheduler
fastapi
uvicorn
requests
python-multipart
fastapi
pytest
